# Optimize ascii.validPrint