--- a/ascii/valid.go
+++ b/ascii/valid.go
@@ -1,3 +1,4 @@
+//go:generate go run valid_asm.go -out valid_amd64.s -stubs valid_amd64.go
 package ascii
 
 import (
@@ -61,12 +62,12 @@ func valid(s unsafe.Pointer, n uintptr) bool {
 
 // Valid returns true if b contains only printable ASCII characters.
 func ValidPrint(b []byte) bool {
-	return validPrint(unsafe.Pointer(&b), uintptr(len(b)))
+	return validPrint(*(*string)(unsafe.Pointer(&b))) != 0
 }
 
 // ValidString returns true if s contains only printable ASCII characters.
 func ValidPrintString(s string) bool {
-	return validPrint(unsafe.Pointer(&s), uintptr(len(s)))
+	return validPrint(s) != 0
 }
 
 // ValidBytes returns true if b is an ASCII character.
@@ -77,45 +78,6 @@ func ValidPrintByte(b byte) bool {
 // ValidBytes returns true if b is an ASCII character.
 func ValidPrintRune(r rune) bool {
 	return 0x20 <= r && r <= 0x7e
-}
-
-//go:nosplit
-func validPrint(s unsafe.Pointer, n uintptr) bool {
-	if n == 0 {
-		return true
-	}
-
-	i := uintptr(0)
-	p := *(*unsafe.Pointer)(s)
-
-	for (n - i) >= 8 {
-		x := *(*uint64)(unsafe.Pointer(uintptr(p) + i))
-		if hasLess64(x, 0x20) || hasMore64(x, 0x7e) {
-			return false
-		}
-		i += 8
-	}
-
-	if (n - i) >= 4 {
-		x := *(*uint32)(unsafe.Pointer(uintptr(p) + i))
-		if hasLess32(x, 0x20) || hasMore32(x, 0x7e) {
-			return false
-		}
-		i += 4
-	}
-
-	var x uint32
-	switch n - i {
-	case 3:
-		x = 0x20000000 | uint32(*(*uint8)(unsafe.Pointer(uintptr(p) + i))) | uint32(*(*uint16)(unsafe.Pointer(uintptr(p) + i + 1)))<<8
-	case 2:
-		x = 0x20200000 | uint32(*(*uint16)(unsafe.Pointer(uintptr(p) + i)))
-	case 1:
-		x = 0x20202000 | uint32(*(*uint8)(unsafe.Pointer(uintptr(p) + i)))
-	default:
-		return true
-	}
-	return !(hasLess32(x, 0x20) || hasMore32(x, 0x7e))
 }
 
 // https://graphics.stanford.edu/~seander/bithacks.html#HasLessInWord