… ascii/valid_default.go
… ascii/valid_test.go
… go.mod
… go.sum
… json/fuzz/fuzz.go