This PR is the result of an experiment I ran, converting the `ascii.validPrint` implementation from Go to assembly. It turned out that the hybrid approach of using Go for short strings, and delegating to the assembly version (using SSE instructions) for longer strings seems to provide the best results; otherwise, the pure assembly implementation was somehow ~20% slower on small strings (I didn't dig deep into why, we're already in good-enough land for short strings).

This optimization is especially useful in cases where we are decoding strings that are base64 encoded (usually going in a `[]byte` field), and tend to embed opaque and lengthy data.

```
name                                old time/op    new time/op     delta
ValidPrint/#00                        2.08ns ± 4%     2.02ns ± 3%     ~     (p=0.063 n=5+5)
ValidPrint/hello                      6.30ns ± 2%     5.70ns ± 1%   -9.62%  (p=0.008 n=5+5)
ValidPrint/Hello_World!               6.63ns ± 3%     6.16ns ± 2%   -7.12%  (p=0.008 n=5+5)
ValidPrint/H~llo_World!               6.52ns ± 2%     6.10ns ± 2%   -6.53%  (p=0.008 n=5+5)
ValidPrint/H~llo                      6.34ns ± 1%     5.64ns ± 2%  -11.16%  (p=0.008 n=5+5)
ValidPrint/你好                         5.81ns ± 3%     5.19ns ± 5%  -10.71%  (p=0.008 n=5+5)
ValidPrint/~                          6.17ns ± 2%     5.07ns ± 2%  -17.88%  (p=0.008 n=5+5)
ValidPrint/�                          6.30ns ± 2%     5.08ns ± 3%  -19.28%  (p=0.008 n=5+5)
ValidPrint/\u007f                     6.13ns ± 2%     5.22ns ± 1%  -14.82%  (p=0.008 n=5+5)
ValidPrint/�#01                       5.93ns ± 1%     5.17ns ± 2%  -12.75%  (p=0.008 n=5+5)
ValidPrint/some_kind_of_long...       13.9ns ± 3%     13.0ns ± 4%   -6.47%  (p=0.016 n=5+5)
ValidPrint/some_kind_of_long...#01    13.5ns ± 1%     10.2ns ± 0%  -24.22%  (p=0.000 n=5+4)
ValidPrint/12345678901234567...       1.13µs ± 1%     0.69µs ± 2%  -38.60%  (p=0.008 n=5+5)

name                                old speed      new speed       delta
ValidPrint/hello                     793MB/s ± 2%    878MB/s ± 1%  +10.67%  (p=0.008 n=5+5)
ValidPrint/Hello_World!             1.81GB/s ± 3%   1.95GB/s ± 2%   +7.66%  (p=0.008 n=5+5)
ValidPrint/H~llo_World!             1.84GB/s ± 2%   1.97GB/s ± 2%   +7.00%  (p=0.008 n=5+5)
ValidPrint/H~llo                     789MB/s ± 2%    887MB/s ± 2%  +12.50%  (p=0.008 n=5+5)
ValidPrint/你好                       1.03GB/s ± 3%   1.16GB/s ± 5%  +12.06%  (p=0.008 n=5+5)
ValidPrint/~                         162MB/s ± 3%    197MB/s ± 2%  +21.75%  (p=0.008 n=5+5)
ValidPrint/�                         159MB/s ± 2%    197MB/s ± 3%  +23.85%  (p=0.008 n=5+5)
ValidPrint/\u007f                    163MB/s ± 2%    191MB/s ± 1%  +17.30%  (p=0.008 n=5+5)
ValidPrint/�#01                      169MB/s ± 1%    193MB/s ± 2%  +14.65%  (p=0.008 n=5+5)
ValidPrint/some_kind_of_long...     3.74GB/s ± 3%   4.00GB/s ± 4%   +6.91%  (p=0.016 n=5+5)
ValidPrint/some_kind_of_long...#01  4.75GB/s ± 1%   6.27GB/s ± 1%  +31.88%  (p=0.008 n=5+5)
ValidPrint/12345678901234567...     8.87GB/s ± 2%  14.45GB/s ± 2%  +62.93%  (p=0.008 n=5+5)
```